+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## June 18, 2026
+
+- `analyze`: add `Token::byte_range` to recover a token's raw source substring
+
 ## May 8, 2026
 
 - `analyze` module, support for stopword removal, stemming, ascii_folding, maximum_token_length, case sensitivity